fix: logs

+++ packages/valory/skills/task_submission_abci/models.py
@@ -76,7 +76,7 @@ class Params(BaseParams):
         )
         self.minimum_agent_balance = self._ensure("minimum_agent_balance", kwargs, int)
         self.agent_funding_amount = self._ensure("agent_funding_amount", kwargs, int)
-        self.polling_interval = self._ensure("polling_interval", kwargs, int)
+        self.polling_interval = self._ensure("polling_interval", kwargs, float)
         super().__init__(*args, **kwargs)
 
     @classmethod
